do not apply the config filter on the
indicators in the sort menu because we want to use it to
select the indicators for the results in the IO Grid. We
will also need checkboxes for the indicators then because
we can then select multiples of them.
