+++ src/widgets/iogrid/commodity-list.tsx
@@ -25,10 +25,11 @@ import {
 import { Indicator, Sector } from "../../webapi";
 import { Config } from "../../widget";
 import { IOGrid } from "./iogrid";
-import { ifNone, isNotNone, TMap } from "../../util/util";
+import { ifNone, TMap } from "../../util/util";
 import * as strings from "../../util/strings";
 
 import { Commodity, SortOptions } from "./commodity-model";
+import { sortIndicators } from "./indicators";
 
 
 const IndicatorValue = new Intl.NumberFormat("en-US", {
@@ -281,9 +282,7 @@ export const CommodityList = (props: {
                                     props.widget.fireChange({ page: 1 });
                                 }
                             }}
-                            indicators={
-                                filterIndicators(props.indicators, props.config)
-                            }
+                            indicators={sortIndicators(props.indicators)}
                             widget={props.widget} />
                     </Menu>
                 </div>
@@ -415,47 +414,3 @@ const SortMenu = React.forwardRef((props: {
 
     return <>{items}</>;
 });
-
-/**
- * Sort the indicators for the `sort-by` menu. We also filter the indicators
- * when a possible indicator filter is set in the configuration here.
- */
-const filterIndicators = (indicators: Indicator[], config: Config): Indicator[] => {
-    if (!indicators) {
-        return [];
-    }
-    const codes = config?.indicators;
-    const filtered = !codes
-        ? indicators.slice(0)
-        : indicators.filter(i => strings.eq(i.code, ...codes));
-    if (!filtered) {
-        return [];
-    }
-
-    // it was specified, that these indicators should be always
-    // at the top of the list ...
-    const predef: TMap<number> = {
-        "Jobs Supported": 1,
-        "Value Added": 2,
-        "Energy Use": 3,
-        "Land Use": 4,
-        "Water Use": 5,
-    };
-
-    return filtered.sort((i1, i2) => {
-        const name1 = i1.simplename || i1.name;
-        const name2 = i2.simplename || i2.name;
-        const c1 = predef[name1];
-        const c2 = predef[name2];
-        if (isNotNone(c1) && isNotNone(c2)) {
-            return c1 - c2;
-        }
-        if (isNotNone(c1)) {
-            return -1;
-        }
-        if (isNotNone(c2)) {
-            return 1;
-        }
-        return strings.compare(name1, name2);
-    });
-};
